Step4

+++ src/main/java/nextstep/courses/domain/SessionRepository.java
@@ -3,7 +3,7 @@ package nextstep.courses.domain;
 import java.util.Optional;
 
 public interface SessionRepository {
-    int save(Session session);
+    long save(Session session);
 
     Optional<Session> findById(Long id);
 }
